update adressing additional PR comments
