+++ DataFormats/EgammaReco/interface/EleSeedSoA.h
@@ -12,15 +12,15 @@ namespace reco {
 
   GENERATE_SOA_LAYOUT(EleSeedLayout,
                       SOA_COLUMN(int8_t, nHits),
-                      SOA_COLUMN(int16_t, isMatched),
-                      SOA_COLUMN(int16_t, matchedScID),
-                      SOA_COLUMN(int16_t, id),
-                      SOA_COLUMN(int16_t, hit0detectorID),
-                      SOA_COLUMN(int16_t, hit0isValid),
-                      SOA_COLUMN(int16_t, hit1detectorID),
-                      SOA_COLUMN(int16_t, hit1isValid),
-                      SOA_COLUMN(int16_t, hit2detectorID),
-                      SOA_COLUMN(int16_t, hit2isValid),
+                      SOA_COLUMN(int8_t, isMatched),
+                      SOA_COLUMN(int8_t, matchedScID),
+                      SOA_COLUMN(int8_t, id),
+                      SOA_COLUMN(int8_t, hit0detectorID),
+                      SOA_COLUMN(int8_t, hit0isValid),
+                      SOA_COLUMN(int8_t, hit1detectorID),
+                      SOA_COLUMN(int8_t, hit1isValid),
+                      SOA_COLUMN(int8_t, hit2detectorID),
+                      SOA_COLUMN(int8_t, hit2isValid),
                       SOA_EIGEN_COLUMN(Vector3d, hit0Pos),
                       SOA_EIGEN_COLUMN(Vector3d, surf0Pos),
                       SOA_EIGEN_COLUMN(Vector3d, surf0Rot),
